+++ src/components/nvi/NviPageLayout.tsx
@@ -1,12 +1,12 @@
 import { Box, Typography } from '@mui/material';
 import { ReactNode } from 'react';
-import { NviInstitutionSearch } from '../../../components/nvi/filters/NviInstitutionSearch';
-import { NviSectorSelector } from '../../../components/nvi/filters/NviSectorSelector';
-import { NviVisibilitySelector } from '../../../components/nvi/filters/NviVisibilitySelector';
-import { NviYearSelector } from '../../../components/nvi/filters/NviYearSelector';
-import { VerticalBox } from '../../../components/styled/Wrappers';
-import { ExportNviStatusButton } from './ExportNviStatusButton';
-import { ExportNviPublicationPointsButton } from '../../../components/nvi/export-buttons/ExportNviPublicationPointsButton';
+import { VerticalBox } from '../styled/Wrappers';
+import { ExportNviPublicationPointsButton } from './export-buttons/ExportNviPublicationPointsButton';
+import { ExportNviStatusButton } from './export-buttons/ExportNviStatusButton';
+import { NviInstitutionSearch } from './filters/NviInstitutionSearch';
+import { NviSectorSelector } from './filters/NviSectorSelector';
+import { NviVisibilitySelector } from './filters/NviVisibilitySelector';
+import { NviYearSelector } from './filters/NviYearSelector';
 
 interface NviStatusWrapperProps {
   headline: string;
@@ -20,7 +20,7 @@ interface NviStatusWrapperProps {
   children?: ReactNode;
 }
 
-export const NviStatusWrapper = ({
+export const NviPageLayout = ({
   headline,
   topView,
   yearSelector,
